Refactor.
Put all the backend code in a "core" file instead of spread throughout
multiple files. Cleanup image crediting code. Remove obsolete, commented-
out bits of code.
