+++ .htaccess
@@ -1,5 +1,5 @@
 RewriteEngine on
 RewriteCond %{REQUEST_FILENAME} !-f
 RewriteCond %{REQUEST_FILENAME} !-d
-RewriteRule ^([^/]*)/([^/]+)/?$ index.php?id=$1&lang=$2 [L,QSA]
-RewriteRule ^([^/]*)/?$ index.php?id=$1 [L,QSA]
+RewriteRule ^(Q[^/]*)/([^/]+)/?$ index.php?id=$1&lang=$2 [L,QSA]
+RewriteRule ^(Q[^/]*)/?$ index.php?id=$1 [L,QSA]
