+++ index.php
@@ -1,15 +1,8 @@
 <?php
-require "inc/helpers.php";
 require "inc/external/getDefaultLanguage.php";
 require "inc/formatting.php";
+require "inc/core.php";
 
-$protocol = ( htmlspecialchars($_SERVER['HTTP_HOST']) == 'localhost' ) ? "http://" : "https://";
-$self = $protocol . htmlspecialchars($_SERVER["HTTP_HOST"]) . "/portal";
-// item id and lang code from url query params
-$item_id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_STRING);
-$lang_code = getBaseLanguage(filter_input(INPUT_GET, 'lang', FILTER_SANITIZE_STRING) ?: getDefaultLanguage());
-$page_title = filter_input(INPUT_GET, 'title', FILTER_SANITIZE_STRING, FILTER_FLAG_NO_ENCODE_QUOTES);
-$page_site = filter_input(INPUT_GET, 'site', FILTER_SANITIZE_STRING);
 
 // When there's no valid item id, nor a page title & site, show the about page instead
 $has_title_and_site = !!$page_title && !! $page_site;
@@ -21,9 +14,6 @@ if ( !preg_match("/^Q\d+$/", $item_id) && !$has_title_and_site ) {
 	echo "</body></html>";
 	die();
 }
-
-require "inc/sites.php";
-require "inc/api.php";
 
 // Backwards compatibility, when given page title & site instead of id
 if ( !$item_id && $has_title_and_site ) {
