+++ web/portal.php
@@ -1,16 +1,21 @@
 <?php
 
-require "inc/items.php";
+$portal = $getPortalInfo();
 
-echo_html_top($item_label);
+echo_html_top($portal["item_label"]);
 
 echo makeHeading(
-	$item_label . makeLangSelector($item_id, "{$self}/img/language_selection.png", $lang_code, $available_langs)
+	$portal["item_label"] . makeLangSelector(
+		$item_id,
+		"{$self}/img/language_selection.png",
+		$lang_code,
+		$available_langs
+	)
 );
-echo makeSubheading($item_desc);
+echo makeSubheading($portal["item_desc"]);
 
 echo "<div class='flex-grid'>";
-foreach ($sitelinks as $site => $site_info) {
+foreach ($portal["sitelinks"] as $site => $site_info) {
 	$sitetype = $site_types[$site];
 	echo makeBoxlink(
 		$site_info["url"],
@@ -21,10 +26,10 @@ foreach ($sitelinks as $site => $site_info) {
 }
 echo "</div>";
 
-if ( count($related_items) > 0 ) {
+if ( count($portal["related_items"]) > 0 ) {
 	echo makeSubheading( getDeepData($i18n, ['related'], 'Related') );	
 	echo "<div class='flex-grid'>";
-	foreach ($related_items as $r) {
+	foreach ($portal["related_items"] as $r) {
 		echo makeBoxlink(
 			"{$self}/{$r['item']}/{$lang_code}",
 			false,
@@ -35,10 +40,10 @@ if ( count($related_items) > 0 ) {
 	echo "</div>";
 }
 
-if ( count($nearby_items) > 0 ) {
+if ( count($portal["nearby_items"]) > 0 ) {
 	echo makeSubheading( getDeepData($i18n, ['nearby'], 'Nearby') );
 	echo "<div class='flex-grid'>";	
-	foreach ($nearby_items as $n) {
+	foreach ($portal["nearby_items"] as $n) {
 		echo makeBoxlink(
 			"{$self}/{$n['item']}/{$lang_code}",
 			false,
@@ -49,8 +54,7 @@ if ( count($nearby_items) > 0 ) {
 	echo "</div>";
 }
 
-$sites_used = array_flip(array_map($mapToSiteType, array_keys($sitelinks)));
-echo makefooter($item_id, $sites_used);
+echo makefooter($item_id, $portal["sites_linked"]);
 
 ?>
  </body>
